Begin implementing Gaussian smoothing function
Minor note: added more details to the README.

This update adds the utils.py file, which will house any code that I
expect will be reused in multiple applications. It begins the
smooth_gauss function, which will apply a Gaussian smoothing filter to a
specified image when complete. Currently it generates a kernel, but does
not do any actual convolution.